moved BulletTest to GdxTests, removed old test
